Connect stdin to /dev/null when tty=False (backwards incompatible*)
Recently I ran into several external commands whose output was being
captured and thus not visible, but which nevertheless rendered an
interactive prompt, waiting for a response on standard input (which
I wasn't providing because I never saw the interactive prompt :-).

The option to connect stdin and /dev/null was never available in
executor, however given the recent addition of the `tty' option it
seemed logical to combine the two.

Two changes in this commit backwards incompatible:

1. The standard input stream of external commands was never connected to
   /dev/null before and this is changing without an explicit opt-in or
   opt-out mechanism. I'm making this choice because I believe it to be
   the only sane approach.

2. The interface of the CachedStream class has changed even though this
   is a documented, externally available class. However I don't actually
   see anyone using CachedStream outside of the executor project, so in
   the grand scheme of things this is a minor thing (99% of users will
   never even notice, I'm guessing).