Trigger backwards incompatibility in test suite
Following up on b4488eea5a6: While it's clear that ExternalCommand.async
needed to be renamed I don't feel like breaking backwards compatibility
with lots of existing Python 2 code using executor with the old 'async'
naming. I'm now adding a test that fails because of the incompatibility.

For more details refer to the following
issues and pull requests on GitHub:

 - https://github.com/xolox/python-executor/issues/9
 - https://github.com/xolox/python-executor/pull/11
 - https://github.com/xolox/python-executor/pull/13