Avoid cyclic dependencies in tcp module
I'm working on readying a new project for publication and started using
the EphemeralTCPServer class in its test suite, however I ran into an
unnecessary cyclic dependency that caused logging to print dramatic
tracebacks (although nothing actually failed):

The __init__() of the EphemeralTCPServer subclass needed access to the
ephemeral TCP port (because it is passed to the command that's started)
which implies running a whole lot of code (to pick a port that isn't in
use yet) and this code logs the EphemeralTCPServer.__str__().

This was intended to use WaitUntilConnected.__str__() however due to
incorrect superclass ordering it called ExternalCommand.__str__()
instead which needs access to the 'command' property which in turn
requires ExternalCommand.__init__() to have already been run.

This catch-22 was broken by removing the __str__() from logging and
using a more explicit newly added 'endpoint' property instead.