Uncover Python 2.7 inconsistency in test suite
It turns out that while shlex.split() on Python 2.7 accepts Unicode
strings it doesn't actually support them: The result is a list of byte
strings and any values that ASCII doesn't support result in an error.

This causes the values (but not the keys) in the dictionary provided by
'lsb_release_variables' to become byte strings which in turn causes
'distributor_id' and 'distribution_codename' to become byte strings
(when those properties are based on 'lsb_release_variables').

However when 'distributor_id' and 'distribution_codename' are based on
the output of the 'lsb_release' program the values become Unicode
strings, so this unfortunate behavior of shlex.split() is causing very
inconsistent behavior in my other carefully designed Python API 😕.