Remove pipes module on versions > 3.3.
Pipes is deprecated as of [3.11](https://peps.python.org/pep-0594/) and
is set to be removed in 3.13.

The `shlex` module introduced a `quote` function in 3.3 and the `pipes`
module adopted it at the same time.

I performed the import in such a way as to continue to use the
pipes.quote when it wasn't just a wrapper around shlex.quote.